refactor: gdal info and standardise-validate
