+++ scripts/standardise_validate.py
@@ -3,17 +3,21 @@ import argparse
 from linz_logger import get_log
 
 from scripts.cli.cli_helper import format_source, is_argo
-from scripts.non_visual_qa import non_visual_qa
+from scripts.files.files_helper import is_tiff
+from scripts.gdal.gdalinfo import gdal_info
+from scripts.non_visual_qa import get_srs, qa_file
 from scripts.standardising import start_standardising
 
 
 def main() -> None:
 
     concurrency: int = 1
-    parse_args = argparse.ArgumentParser()
-    parse_args.add_argument("--preset", dest="preset", required=True)
-    parse_args.add_argument("--source", dest="source", nargs="+", required=True)
-    arguments = parse_args.parse_args()
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--preset", dest="preset", required=True)
+    parser.add_argument("--source", dest="source", nargs="+", required=True)
+
+    arguments = parser.parse_args()
 
     source = format_source(arguments.source)
 
@@ -21,12 +25,17 @@ def main() -> None:
         concurrency = 4
 
     standardised_files = start_standardising(source, arguments.preset, concurrency)
-    if standardised_files:
-        non_visual_qa(standardised_files)
-    else:
-        get_log().info(
-            "Non Visual QA skipped because no file has been standardised", action="standardise_validate", reason="skip"
-        )
+    if not standardised_files:
+        get_log().info("Process skipped because no file has been standardised", action="standardise_validate", reason="skip")
+        return
+    srs = get_srs()
+
+    for file in standardised_files:
+        if not is_tiff(file):
+            get_log().trace("file_not_tiff_skipped", file=file)
+            continue
+        gdalinfo_result = gdal_info(file)
+        qa_file(file, srs, gdalinfo_result)
 
 
 if __name__ == "__main__":
